Openapi wines (#78)
Use OpenAPI for wine endpoints

+++ src/main/java/com/mika/WineApp/services/CrudService.java
@@ -4,17 +4,19 @@ import com.mika.WineApp.errors.notfound.NotFoundException;
 
 import java.util.List;
 
-public interface CrudService <EntityModel> {
+public interface CrudService<EntityModel> {
     // Add operation is missing because it is different for wines and reviews.
 
     /**
      * Get all wines or reviews.
+     *
      * @return list of entities
      */
     List<EntityModel> findAll();
 
     /**
      * Find wine or review by ID.
+     *
      * @param id of entity
      * @return wine or review
      * @throws NotFoundException e
@@ -23,6 +25,7 @@ public interface CrudService <EntityModel> {
 
     /**
      * Edit wine or review.
+     *
      * @param id of persisted entity to edit
      * @param em edited wine or review
      * @return persisted entity model
@@ -32,13 +35,15 @@ public interface CrudService <EntityModel> {
 
     /**
      * Delete wine or review.
+     *
      * @param id of entity
      */
     void delete(Long id);
 
     /**
      * Get entity count in database.
+     *
      * @return Long count.
      */
-    long count();
+    int count();
 }
